Working on 'main', batcher is working great with a few small issues.
The weaken threads didn't calc the time themselves and always used the start time,
which isn't accurate at all if the time changes.   I could communicate with a file
becaue reading a file is free, or simply put the time on a different port and have
them peek at it?  Also the config I used I don't think works because the weaken threads
can't handle the grows, stupid analyzer...